refactor(form): split up input/field/item resolvers and move renderItem to config level
