+++ packages/sanity/src/core/form/inputs/arrays/ArrayOfObjectsInput/Grid/GridItem.tsx
@@ -10,7 +10,7 @@ import {
   Spinner,
   Text,
 } from '@sanity/ui'
-import React, {ReactNode, useCallback, useMemo, useRef} from 'react'
+import React, {useCallback, useMemo, useRef} from 'react'
 import {SchemaType} from '@sanity/types'
 import {CopyIcon as DuplicateIcon, EllipsisVerticalIcon, TrashIcon} from '@sanity/icons'
 import styled from 'styled-components'
@@ -30,12 +30,8 @@ import {InsertMenu} from '../InsertMenu'
 import {FIXME} from '../../../../../FIXME'
 import {EditPortal} from '../../../../components/EditPortal'
 
-interface Props<Item extends ObjectItem> extends Omit<ObjectItemProps<Item>, 'renderDefault'> {
-  insertableTypes: SchemaType[]
-  value: Item
-  preview: ReactNode
-  sortable: boolean
-}
+type GridItemProps<Item extends ObjectItem> = Omit<ObjectItemProps<Item>, 'renderDefault'>
+
 const PreviewCard = styled(Card)`
   border-top-right-radius: inherit;
   border-top-left-radius: inherit;
@@ -79,9 +75,10 @@ const INITIAL_VALUE_CARD_STYLE = {
   right: 0,
 } as const
 
-export function GridItem<Item extends ObjectItem = ObjectItem>(props: Props<Item>) {
+export function GridItem<Item extends ObjectItem = ObjectItem>(props: GridItemProps<Item>) {
   const {
     schemaType,
+    parentSchemaType,
     path,
     readOnly,
     onRemove,
@@ -91,14 +88,14 @@ export function GridItem<Item extends ObjectItem = ObjectItem>(props: Props<Item
     onFocus,
     onOpen,
     onClose,
-    inputId,
     changed,
     focused,
     children,
-    sortable,
-    preview,
-    insertableTypes,
+    inputProps: {renderPreview},
   } = props
+
+  const sortable = parentSchemaType.options?.sortable !== false
+  const insertableTypes = parentSchemaType.of
 
   const previewCardRef = useRef<FIXME | null>(null)
 
@@ -198,7 +195,11 @@ export function GridItem<Item extends ObjectItem = ObjectItem>(props: Props<Item
         onFocus={onFocus}
         __unstable_focusRing
       >
-        {preview}
+        {renderPreview({
+          schemaType,
+          value,
+          layout: 'media',
+        })}
         {resolvingInitialValue && (
           <Card as={Flex} style={INITIAL_VALUE_CARD_STYLE}>
             <Flex align="center" justify="center" gap={1} padding={1}>
