+++ packages/sanity/src/core/form/members/items/ArrayOfObjectsItem.tsx
@@ -239,6 +239,7 @@ export function ArrayOfObjectsItem(props: MemberItemProps) {
       collapsible: member.collapsible,
       collapsed: member.collapsed,
       schemaType: member.item.schemaType,
+      parentSchemaType: member.parentSchemaType,
       onInsert,
       onRemove,
       presence: member.item.presence,
@@ -264,6 +265,7 @@ export function ArrayOfObjectsItem(props: MemberItemProps) {
     member.item.level,
     member.item.value,
     member.item.schemaType,
+    member.parentSchemaType,
     member.item.presence,
     member.item.validation,
     member.item.readOnly,
