+++ packages/sanity/src/core/form/members/items/ArrayOfPrimitivesItem.tsx
@@ -156,6 +156,7 @@ export function ArrayOfPrimitivesItem(props: PrimitiveMemberItemProps) {
       title: member.item.schemaType.title,
       description: member.item.schemaType.description,
       schemaType: member.item.schemaType as FIXME,
+      parentSchemaType: member.parentSchemaType,
       onInsert,
       onRemove,
       presence: member.item.presence,
@@ -180,6 +181,7 @@ export function ArrayOfPrimitivesItem(props: PrimitiveMemberItemProps) {
     member.item.focused,
     member.item.id,
     member.item.path,
+    member.parentSchemaType,
     onInsert,
     onRemove,
     handleFocus,
